Remote Managaing Your VMS
A script with which you will be able to manage your servers remotely. It is important to enable ps-remoting on each server. The script includes one function to enable the virtual machine - you can add other functions to make the script more interesting.
I think this is a good basis

:)